refactored context to be specific to dao creation
